+++ spec/open_meteo/forecast/variables_spec.rb
@@ -1,8 +1,9 @@
 RSpec.describe OpenMeteo::Forecast::Variables do
-  let(:variables) { described_class.new(current:, hourly:, daily:) }
+  let(:variables) { described_class.new(current:, hourly:, daily:, models:) }
   let(:current) { [] }
   let(:hourly) { [] }
   let(:daily) { [] }
+  let(:models) { [] }
 
   describe "#to_get_params" do
     subject { variables.to_get_params }
@@ -11,6 +12,7 @@ RSpec.describe OpenMeteo::Forecast::Variables do
       let(:current) { [] }
       let(:hourly) { [] }
       let(:daily) { [] }
+      let(:models) { [] }
 
       it { is_expected.to eq({}) }
     end
@@ -19,6 +21,7 @@ RSpec.describe OpenMeteo::Forecast::Variables do
       let(:current) { %i[something other] }
       let(:hourly) { [] }
       let(:daily) { [] }
+      let(:models) { [] }
 
       it { is_expected.to eq({ current: "something,other" }) }
     end
@@ -27,6 +30,7 @@ RSpec.describe OpenMeteo::Forecast::Variables do
       let(:current) { [] }
       let(:hourly) { %i[something other] }
       let(:daily) { [] }
+      let(:models) { [] }
 
       it { is_expected.to eq({ hourly: "something,other" }) }
     end
@@ -35,16 +39,32 @@ RSpec.describe OpenMeteo::Forecast::Variables do
       let(:current) { [] }
       let(:hourly) { [] }
       let(:daily) { %i[something other] }
+      let(:models) { [] }
 
       it { is_expected.to eq({ daily: "something,other" }) }
+    end
+
+    context "when models is set" do
+      let(:current) { [] }
+      let(:hourly) { [] }
+      let(:daily) { [] }
+      let(:models) { %i[something other] }
+
+      it { is_expected.to eq({ models: "something,other" }) }
     end
 
     context "when all are set" do
       let(:current) { %i[something other] }
       let(:hourly) { %i[something other] }
       let(:daily) { %i[something other] }
+      let(:models) { %i[something other] }
       let(:expected_output) do
-        { current: "something,other", hourly: "something,other", daily: "something,other" }
+        {
+          current: "something,other",
+          hourly: "something,other",
+          daily: "something,other",
+          models: "something,other",
+        }
       end
 
       it { is_expected.to eq(expected_output) }