+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Add retry logic to the client
+- Add models to forecast variables
 
 ## [0.2.0] - 2023-11-08